Shuffle things to avoid circular dep
